+++ packages/koa-openapi/index.ts
@@ -109,7 +109,7 @@ export function initialize(args: KoaOpenAPIInitializeArgs): OpenAPIFramework {
       if (operationDoc && operationCtx.allowsFeatures) {
         if (operationCtx.features.requestValidator) {
           middleware.unshift(function requestValidatorMiddleware(ctx: Context) {
-            const errors = operationCtx.features.requestValidator.validate(
+            const errors = operationCtx.features.requestValidator.validateRequest(
               toOpenAPIRequest(ctx)
             );
             if (errors) {