+++ packages/openapi-response-validator/test/data-driven/fail-response-against-undefined-schema.js
@@ -12,8 +12,9 @@ module.exports = {
     message: 'The response was not valid.',
     errors: [
       {
+        path: 'response',
         errorCode: 'type.openapi.responseValidation',
-        message: 'response should be null'
+        message: 'should be null'
       }
     ]
   }