+++ packages/openapi-response-validator/test/data-driven/fail-with-external-references-through-local-ref.js
@@ -35,7 +35,7 @@ module.exports = {
       {
         path: 'foo',
         errorCode: 'type.openapi.responseValidation',
-        message: 'foo should be string'
+        message: 'should be string'
       }
     ]
   }