+++ packages/openapi-jsonschema-parameters/CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 1.2.0 - 2019-10-21
+### Aded
+- Added support for 'examples' (fixes #513) (#514)
+
+### Fixed
+- remove extra parens from example code (#550)
+
 ## 1.1.2 - 2019-06-21
 ### Fixed
 - Adds example keyword (fixes #455) (#456)